smartHome/requirements.txt
paho-mqtt==1.6.1
PyQt5==5.15.11
